updated tagview, issue-7
